+++ config_bayesian.py
@@ -1,15 +1,15 @@
 ############### Configuration file for Bayesian ###############
-layer_type = 'lrt'
-activation_type = 'softplus'
+layer_type = 'lrt'  # 'bbb' or 'lrt'
+activation_type = 'softplus'  # 'softplus' or 'relu'
 
-n_epochs = 100
+n_epochs = 5
 lr_start = 0.001
 num_workers = 4
 valid_size = 0.2
 batch_size = 256
 train_ens = 1
 valid_ens = 1
-beta_type = 0.2  # 'Blundell', 'Standard', etc. Use float for const value
+beta_type = 0.1  # 'Blundell', 'Standard', etc. Use float for const value
 
 record_mean_var = False
 recording_freq_per_epoch = 32
