+++ layers/BBB/BBBLinear.py
@@ -7,14 +7,11 @@ import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn import Parameter
 
-import utils
 from metrics import calculate_kl as KL_DIV
-import config_bayesian as cfg
 from ..misc import ModuleWrapper
 
 
 class BBBLinear(ModuleWrapper):
-    
     def __init__(self, in_features, out_features, bias=True):
         super(BBBLinear, self).__init__()
         self.in_features = in_features
@@ -25,11 +22,12 @@ class BBBLinear(ModuleWrapper):
         self.prior_mu = 0
         self.prior_sigma = 0.1
 
-        self.W_mu = Parameter(torch.Tensor(out_features, in_features))
-        self.W_rho = Parameter(torch.Tensor(out_features, in_features))
+        self.W_mu = Parameter(torch.empty((out_features, in_features), device=self.device))
+        self.W_rho = Parameter(torch.empty((out_features, in_features), device=self.device))
+
         if self.use_bias:
-            self.bias_mu = Parameter(torch.Tensor(out_features))
-            self.bias_rho = Parameter(torch.Tensor(out_features))
+            self.bias_mu = Parameter(torch.empty((out_features), device=self.device))
+            self.bias_rho = Parameter(torch.empty((out_features), device=self.device))
         else:
             self.register_parameter('bias_mu', None)
             self.register_parameter('bias_rho', None)
@@ -44,24 +42,23 @@ class BBBLinear(ModuleWrapper):
             self.bias_mu.data.normal_(0, 0.1)
             self.bias_rho.data.normal_(-3, 0.1)
 
-    def forward(self, x, sample=True):
-
-        self.W_sigma = torch.log1p(torch.exp(self.W_rho))
-        if self.use_bias:
-            self.bias_sigma = torch.log1p(torch.exp(self.bias_rho))
-            bias_var = self.bias_sigma ** 2
-        else:
-            self.bias_sigma = bias_var = None
-
-        act_mu = F.linear(x, self.W_mu, self.bias_mu)
-        act_var = 1e-16 + F.linear(x ** 2, self.W_sigma ** 2, bias_var)
-        act_std = torch.sqrt(act_var)
-
+    def forward(self, input, sample=True):
         if self.training or sample:
-            eps = torch.empty(act_mu.size()).normal_(0, 1).to(self.device)
-            return act_mu + act_std * eps
+            W_eps = torch.empty(self.W_mu.size()).normal_(0, 1).to(self.device)
+            self.W_sigma = torch.log1p(torch.exp(self.W_rho))
+            weight = self.W_mu + W_eps * self.W_sigma
+
+            if self.use_bias:
+                bias_eps = torch.empty(self.bias_mu.size()).normal_(0, 1).to(self.device)
+                self.bias_sigma = torch.log1p(torch.exp(self.bias_rho))
+                bias = self.bias_mu + bias_eps * self.bias_sigma
+            else:
+                bias = None
         else:
-            return act_mu
+            weight = self.W_mu
+            bias = self.bias_mu if self.use_bias else None
+
+        return F.linear(input, weight, bias)
 
     def kl_loss(self):
         kl = KL_DIV(self.prior_mu, self.prior_sigma, self.W_mu, self.W_sigma)
