+++ layers/__init__.py
@@ -1,7 +1,7 @@
+from .BBB.BBBLinear import BBBLinear as BBB_Linear
+from .BBB.BBBConv import BBBConv2d as BBB_Conv2d
+
 from .BBB_LRT.BBBLinear import BBBLinear as BBB_LRT_Linear
 from .BBB_LRT.BBBConv import BBBConv2d as BBB_LRT_Conv2d
-
-from .BBB_MCMF_LRT.BBBLinear import BBBLinear as BBB_MCMF_LRT_Linear
-from .BBB_MCMF_LRT.BBBConv import BBBConv2d as BBB_MCMF_LRT_Conv2d
 
 from .misc import FlattenLayer, ModuleWrapper
